requirements.txt
config==0.3.9
dj-config-url==0.1.1
Django==1.11.6
djangorestframework==3.7.1
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
python-decouple==3.1
pytz==2017.2
whitenoise==3.3.1
psycopg2==2.7.1
